linear boost controller tests

+++ contracts/core/interfaces/IBoostController.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.8.15;
 
 interface IBoostController {
-    function getBoostAmount(uint256 _amount, uint256 _lockingDuration) external view returns (uint256);
+    function getBoostAmount(uint256 _amount, uint64 _lockingDuration) external view returns (uint256);
 }
